+++ app/src/main/java/com/example/myapplication/database/table/SettingDate.java
@@ -7,18 +7,18 @@ import androidx.room.PrimaryKey;
 public class SettingDate {
 
     @PrimaryKey(autoGenerate = true)
-    private int settingId;      // 0: 오픈, 1: 마감, 2: 시간간격, 3: 휴무
+    private long settingId;      // 0: 오픈, 1: 마감, 2: 시간간격, 3: 휴무
     private String settingTime; // 0~2:시간.                    3: 1월,2화,3수,4목,5금,6토,7일
 
     public SettingDate(String settingTime){
         this.settingTime = settingTime;
     }
 
-    public int getSettingId() {
+    public long getSettingId() {
         return settingId;
     }
 
-    public void setSettingId(int settingId) {
+    public void setSettingId(long settingId) {
         this.settingId = settingId;
     }
